backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose==3.3.0
passlib
bcrypt
python-dotenv
pydantic
alembic
python-multipart
httpx
python-dateutil
apscheduler
